+++ backend/internal/domain/repo/jobs/order.go
@@ -7,10 +7,10 @@ import (
 )
 
 type OrderRepository interface {
-	Create(ctx context.Context, jobOrder *jobs.JobOrder) (int, error)
-	ExistsByOrderID(ctx context.Context, orderId string) int
-	First(ctx context.Context, id int) (*jobs.JobOrder, error)
-	UpdateJobTime(ctx context.Context, id int) error
-	UpdateStatus(ctx context.Context, id int, status int) error
+	Create(ctx context.Context, jobOrder *jobs.JobOrder) (int64, error)
+	ExistsByOrderID(ctx context.Context, orderId string) int64
+	First(ctx context.Context, id int64) (*jobs.JobOrder, error)
+	UpdateJobTime(ctx context.Context, id int64) error
+	UpdateStatus(ctx context.Context, id int64, status int) error
 	Clear(ctx context.Context) error
 }
